refactor: reorganize internal structure and enhance logging capabilities
