+++ packages/kubricate/src/commands/generate.ts
@@ -1,10 +1,10 @@
 import c from 'ansis';
-import { MARK_CHECK, MARK_NODE } from '../constant.js';
-import { getClassName } from '../utils.js';
+import { MARK_CHECK, MARK_NODE } from '../internal/constant.js';
+import { getClassName } from '../internal/utils.js';
 import { stringify as yamlStringify } from 'yaml';
 import fs from 'node:fs/promises';
 import path from 'node:path';
-import type { GlobalConfigOptions } from '../types.js';
+import type { GlobalConfigOptions } from '../internal/types.js';
 import type { BaseLogger } from '@kubricate/core';
 import { BaseCommand } from './base.js';
