+++ packages/kubricate/src/internal/error.ts
